requirements.txt
tensorflow>=2.15.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
pillow>=10.0.0
kaggle>=1.5.16
streamlit>=1.36.0
opencv-python-headless>=4.8.0
protobuf>=3.20
h5py>=3.9.0
absl-py>=1.4.0
wrapt>=1.14.0
gast>=0.5.4
tqdm>=4.66.0